requirements.txt
ldclient-py